# api/api-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IDClawserver API",
    "version": "0.1.0",
    "description": "API for AI agent identities backed by RODiT tokens on NEAR, including facial token_id encoding.",
    "x-policies": {
      "termsOfService": "/.well-known/terms-of-service",
      "privacyPolicy": "/.well-known/privacy-policy",
      "dataRetention": "/.well-known/data-retention"
    },
    "contact": {
      "name": "IDENTYCLAW Support",
      "email": "support@identityclaw.com",
      "url": "https://identityclaw.com/contact"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://identityclaw.com/license"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "RODiT-JWT",
        "description": "RODiT-issued JWT access token validated by @rodit/rodit-auth-be (RoditClient.authenticate), obtained via the RODiT login flow (e.g. /api/login). This is not an API key."
      }
    },
    "schemas": {
...
      "DidVerificationMethod": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier for the verification method fragment."
          },
          "type": {
            "type": "string",
            "description": "Verification method type, e.g. Ed25519VerificationKey2020."
          },
          "controller": {
            "type": "string",
            "description": "Controller DID for the verification method."
          },
          "publicKeyBase58": {
            "type": "string",
            "description": "Base58-encoded Ed25519 public key."
          }
        },
        "required": ["id", "type", "controller", "publicKeyBase58"]
      },
      "DidService": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Service fragment identifier."
          },
          "type": {
            "type": "string",
            "description": "Service type identifier."
          },
          "serviceEndpoint": {
            "description": "Service endpoint location or descriptor.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": ["id", "type", "serviceEndpoint"]
      },
      "DidDocument": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "array",
            "description": "Ordered list of JSON-LD contexts used in the DID document.",
            "items": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "object"
                }
              ]
            }
          },
          "id": {
            "type": "string",
            "description": "Primary DID identifier."
          },
          "alsoKnownAs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Alternate DID identifiers for the same subject."
          },
          "controller": {
            "type": "string",
            "description": "Controller account for the DID document."
          },
          "verificationMethod": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DidVerificationMethod"
            },
            "description": "Verification methods available on the DID document."
          },
          "authentication": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assertionMethod": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DidService"
            },
            "description": "Service endpoints advertised in the DID document."
          },
          "requestId": {
            "type": "string",
            "description": "Request correlation identifier returned by the API."
          }
        },
        "required": ["@context", "id", "controller", "verificationMethod", "authentication", "assertionMethod", "service", "requestId"]
      },
...
    }
  },
  "paths": {
...
    "/api/mcp/resources": {
      "get": {
        "summary": "List available MCP resources (including OpenAPI schema)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available MCP resources."
          },
          "500": {
            "description": "Failed to list resources."
          }
        }
      }
    },
...
    "/.well-known/did/rodit/{tokenId}": {
      "get": {
        "summary": "Resolve did:rodit DID document",
        "description": "Returns the DID document for a specific RODiT token identifier. This endpoint requires authentication because DID resolution is a paid service.",
        "tags": ["DID"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "RODiT token identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "DID document for the requested token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DidDocument"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters."
          },
          "404": {
            "description": "Token not found."
          }
        }
      }
    },
    "/.well-known/did/resolve": {
      "get": {
        "summary": "Resolve DID string",
        "description": "Resolves either did:rodit or did:web identifiers into DID documents.",
        "tags": ["DID"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "did",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "DID string to resolve (did:rodit or did:web forms)."
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved DID document.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DidDocument"
                }
              }
            }
          },
          "400": {
            "description": "Invalid DID supplied."
          },
          "404": {
            "description": "Token not found for the supplied DID."
          }
        }
      }
    },
    "/.well-known/did/web/token/{tokenId}": {
      "get": {
        "summary": "Resolve did:web DID document",
        "description": "Returns the did:web representation for a token hosted by this service, providing interoperability with web DID resolvers.",
        "tags": ["DID"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "RODiT token identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "did:web DID document for the requested token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DidDocument"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters."
          },
          "404": {
            "description": "Token not found."
          }
        }
      }
    },
    "/.well-known/did/web/token/{tokenId}/did.json": {
      "get": {
        "summary": "Resolve did:web JSON DID document",
        "description": "Returns the JSON DID document commonly served at did.json endpoints for did:web identifiers.",
        "tags": ["DID"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "RODiT token identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "did:web DID document for the requested token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DidDocument"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters."
          },
          "404": {
            "description": "Token not found."
          }
        }
      }
    },
...
  }
}